Move ContentInstaller to OpenRA.Game from Mods.Common
